fixed so changepassword is shown again

+++ src/containers/ChangeInfoPage/ChangeInfoPage.js
@@ -45,6 +45,7 @@ class ChangeInfoPage extends Component {
         wantToEdit: false,
         imageData: null,
         imageUrl: this.props.navigation.getParam('imageUrl', null),
+        showImagePicker: this.props.navigation.getParam('showImagePicker', true),
     }
 
     handleInputChange = (value, key) => {
@@ -84,6 +85,17 @@ class ChangeInfoPage extends Component {
             case 'Company Department':
                 formErrors.companyDepartment = value.length < 1 || value.length > 3 ? "Invalid department" : "";
                 break;
+            case 'Password':
+                formErrors.password = value.length < 6 ? "Minimum 6 characters required" : "";
+                formErrors.confirmPassword = value === fields['confirmPassword'].value ? "" : "Passwords must match";
+                break;
+            case 'New Password':
+                formErrors.password = value.length < 6 ? "Minimum 6 characters required" : "";
+                formErrors.confirmPassword = value === fields['confirmPassword'].value ? "" : "Passwords must match";
+                break;
+            case 'Confirm Password':
+                formErrors.confirmPassword = value === fields['password'].value ? "" : "Passwords must match";
+                break;
             default:
                 break;
         }
@@ -99,31 +111,32 @@ class ChangeInfoPage extends Component {
             }, new FormData())
             body.append("title", this.state.title);
 
-            let image = null;
-            if (this.state.imageData) {
-                image = {
-                    name: this.state.imageData.fileName,
-                    type: this.state.imageData.type,
-                    uri:
-                        Platform.OS === "android" ? this.state.imageData.uri : this.state.imageData.uri.replace("file://", "")
+            if (this.state.showImagePicker) {
+                let image = null;
+                if (this.state.imageData) {
+                    image = {
+                        name: this.state.imageData.fileName,
+                        type: this.state.imageData.type,
+                        uri:
+                            Platform.OS === "android" ? this.state.imageData.uri : this.state.imageData.uri.replace("file://", "")
+                    }
+                }
+                body.append("image", image)
+
+                if (this.state.imageData == null) {
+                    let axiosUrl;
+                    if (this.state.parentRoute === "ProfilePageRoute") {
+                        axiosUrl = URL + 'users/' + this.props.userID + '/profileimage';
+                    } else {
+                        //add this line when delete event image works in BE
+                        //axiosUrl = http_update_url + '/coverImage';
+                    }
+                    axios.delete(axiosUrl)
+                        .catch((error) => {
+                            console.log(error.response);
+                        })
                 }
             }
-            body.append("image", image)
-
-            if (this.state.imageData == null) {
-                let axiosUrl;
-                if (this.state.parentRoute === "ProfilePageRoute") {
-                    axiosUrl = URL + 'users/' + this.props.userID + '/profileimage';
-                } else {
-                    //add this line when delete event image works in BE
-                    //axiosUrl = http_update_url + '/coverImage';
-                }
-                axios.delete(axiosUrl)
-                    .catch((error) => {
-                        console.log(error.response);
-                    })
-            }
-
             console.log("body: ", body);
 
             this.setState({ isLoading: true }, () => {
@@ -182,6 +195,7 @@ class ChangeInfoPage extends Component {
     }
 
     render() {
+
         return (
             !this.state ? <View /> :
                 <View style={styles.pageContainer}>
@@ -200,13 +214,17 @@ class ChangeInfoPage extends Component {
                             >{'Edit ' + this.state.title}
                             </HeadlineOverview>
 
-                            <ImageSelector
-                                saveImageHandler={this.saveImageHandler}
-                                deleteImageHandler={this.deleteImageHandler}
-                                parentRoute={this.state.parentRoute}
-                                source={{ uri: this.state.imageUrl }}>
-                                Press to change photo
-                            </ImageSelector>
+                            {
+                                this.state.showImagePicker ?
+                                    <ImageSelector
+                                        saveImageHandler={this.saveImageHandler}
+                                        deleteImageHandler={this.deleteImageHandler}
+                                        parentRoute={this.state.parentRoute}
+                                        source={{ uri: this.state.imageUrl }}>
+                                        Press to change photo
+                                    </ImageSelector> :
+                                    null
+                            }
 
                             <View style={styles.editFormContainer}>
                                 <EditableForm
